+++ app/templates/auth/logout.html
