+++ app/templates/photos/gallery.html
