+++ app/templates/photos/upload.html
