+++ app/templates/resources/pups.html
